requirements.txt
absl-py
aom
appdirs
astroid
astunparse
av
black
blas
blosc
bottleneck
brotli
brotli-bin
brotlipy
bzip2
ca-certificates
cachetools
certifi
cffi
cfgv
cfitsio
charls
charset-normalizer
click
cloudpickle
colorama
contourpy
cryptography
cycler
cytoolz
dask-core
dill
distlib
eigen
expat
ffmpeg
filelock
flask
flatbuffers
flit-core
font-ttf-dejavu-sans-mono
font-ttf-inconsolata
font-ttf-source-code-pro
font-ttf-ubuntu
fontconfig
fonts-anaconda
fonts-conda-ecosystem
fonttools
freetype
fsspec
future
gast
giflib
glib
google-auth
google-auth-oauthlib
google-pasta
grpcio
gst-plugins-base
gstreamer
h5py
hdf5
icu
identify
idna
imagecodecs
imageio
importlib-metadata
imutils
intel-openmp
isort
itsdangerous
jax
jinja2
joblib
jpeg
jxrlib
keras
kiwisolver
lazy-object-proxy
lcms2
lerc
libaec
libbrotlicommon
libbrotlidec
libbrotlienc
libclang
libdeflate
libffi
libiconv
libogg
libpng
libprotobuf
libtiff
libuv
libvorbis
libwebp
libwebp-base
libxml2
libxslt
libzlib
libzopfli
locket
lz4-c
markdown
markupsafe
matplotlib
matplotlib-base
mccabe
mkl
mkl-service
munkres
networkx
ninja
ninja-base
nodeenv
numexpr
numpy
numpy-base
oauthlib
opencv
openh264
openjpeg
openssl
opt-einsum
packaging
pandas
partd
pathspec
pcre
pillow
pip
platformdirs
ply
pooch
pre-commit
protobuf
psycopg2
pyasn1
pyasn1-modules
pycparser
pyjwt
pylint
pyopenssl
pyparsing
pyqt
pyqt5-sip
pysocks
python
python-dateutil
pytorch
pytz
pywavelets
pyyaml
qt-main
qt-webengine
qtwebkit
requests
requests-oauthlib
rsa
scikit-image
scikit-learn
scipy
setuptools
sip
six
snappy
sqlite
svt-av1
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tifffile
tk
toml
tomli
tomlkit
toolz
torchvision
tornado
tqdm
twilio
typing
typing-extensions
ucrt
ukkonen
urllib3
vc
virtualenv
werkzeug
wheel
wincertstore
wrapt
x264
x265
xz
yaml
zfp
zipp
zlib
zstd
